+++ docs/graph/examples/hld/hld_2.ans
@@ -1,0 +1,2 @@
+2
+21
