+++ docs/graph/examples/hld/hld_2.in
@@ -1,0 +1,11 @@
+5 5 2 24
+7 3 7 8 0 
+1 2
+1 5
+3 1
+4 1
+3 4 2
+3 2 2
+4 5
+1 5 1 3
+2 1 3
