+++ docs/graph/hld.md
@@ -290,11 +290,9 @@ $1 \le n, m, \le 10^5$。
 对于此题提供一份完全由指针实现的代码。
 
 ??? note "参考代码"
-
-
-```cpp
-  --8<-- "docs/graph/code/hld/hld_4.cpp"
-```
+    ```cpp
+      --8<-- "docs/graph/code/hld/hld_2.cpp"
+    ```
 
 ### [Nauuo and Binary Tree](https://loj.ac/problem/6669)
 
@@ -337,11 +335,90 @@ $$
 $2999+\sum_{i=1}^{2999}T(i)\le 29940$，事实上这个上界是可以通过构造数据达到的，然而只要进行一些随机扰动（如对深度进行排序时使用不稳定的排序算法），询问次数很难超过 $21000$ 次。
 
 ??? note "参考代码"
+    ```cpp
+    #include <algorithm>
+    #include <cstdio>
+    #include <iostream>
 
+    using namespace std;
 
-```cpp
-  --8<-- "docs/graph/code/hld/hld_2.cpp"
-```
+    const int N = 3010;
+
+    int n, fa[N], ch[N][2], dep[N], siz[N], son[N], bot[N], id[N];
+
+    int query(int u, int v) {
+      printf("? %d %d\n", u, v);
+      fflush(stdout);
+      int d;
+      scanf("%d", &d);
+      return d;
+    }
+
+    void setFather(int u, int v) {
+      fa[v] = u;
+      if (ch[u][0])
+        ch[u][1] = v;
+      else
+        ch[u][0] = v;
+    }
+
+    void dfs(int u) {
+      if (ch[u][0]) dfs(ch[u][0]);
+      if (ch[u][1]) dfs(ch[u][1]);
+
+      siz[u] = siz[ch[u][0]] + siz[ch[u][1]] + 1;
+
+      if (ch[u][1])
+        son[u] = int(siz[ch[u][0]] < siz[ch[u][1]]);
+      else
+        son[u] = 0;
+
+      if (ch[u][son[u]])
+        bot[u] = bot[ch[u][son[u]]];
+      else
+        bot[u] = u;
+    }
+
+    void solve(int u, int k) {
+      if (!ch[u][0]) {
+        setFather(u, k);
+        return;
+      }
+      int d = query(k, bot[u]);
+      int v = bot[u];
+      while (dep[v] > (dep[k] + dep[bot[u]] - d) / 2) v = fa[v];
+      int w = ch[v][son[v] ^ 1];
+      if (w)
+        solve(w, k);
+      else
+        setFather(v, k);
+    }
+
+    int main() {
+      int i;
+
+      scanf("%d", &n);
+
+      for (i = 2; i <= n; ++i) {
+        id[i] = i;
+        dep[i] = query(1, i);
+      }
+
+      sort(id + 2, id + n + 1, [](int x, int y) { return dep[x] < dep[y]; });
+
+      for (i = 2; i <= n; ++i) {
+        dfs(1);
+        solve(1, id[i]);
+      }
+
+      printf("!");
+      for (i = 2; i <= n; ++i) printf(" %d", fa[i]);
+      printf("\n");
+      fflush(stdout);
+
+      return 0;
+    }
+    ```
 
 ## 长链剖分
